Crem 2

+++ ExternalServices/Crem/Dto/FlatsForAddress.cs
@@ -10,7 +10,7 @@ public sealed class FlatsForAddress
     /// <summary>
     /// ISKN ID listu vlastnictví(LV), technický identifikátor katastru
     /// </summary>
-    public int DeedOfOwnershipId { get; set; }
+    public long DeedOfOwnershipId { get; set; }
 
     /// <summary>
     /// Byt
@@ -25,7 +25,7 @@ public sealed class FlatsForAddress
         /// <summary>
         /// ISKN ID listu vlastnictví (LV), technický identifikátor katastru
         /// </summary>
-        public int DeedOfOwnershipId { get; set; }
+        public long DeedOfOwnershipId { get; set; }
 
         /// <summary>
         /// Způsob využití bytu
